+++ OrcamentoApp/OrcamentoApp/Models/Encargos.cs
@@ -14,11 +14,11 @@ namespace OrcamentoApp.Models
     
     public partial class Encargos
     {
-        public double Enc13 { get; set; }
-        public double AvisoPrevio { get; set; }
-        public double Ferias { get; set; }
-        public double FGTS { get; set; }
-        public double INSS { get; set; }
+        public float Enc13 { get; set; }
+        public float AvisoPrevio { get; set; }
+        public float Ferias { get; set; }
+        public float FGTS { get; set; }
+        public float INSS { get; set; }
         public int EmpresaCod { get; set; }
         public float INCRA { get; set; }
         public float SalEducacao { get; set; }
