+++ OrcamentoApp/OrcamentoApp/Models/Reajuste.cs
@@ -18,7 +18,7 @@ namespace OrcamentoApp.Models
         public double PercentualReajuste { get; set; }
         public int MesFechamento { get; set; }
         public int MesReajuste { get; set; }
-        public double PisoSalarial { get; set; }
+        public float PisoSalarial { get; set; }
         public int SindicatoCod { get; set; }
     
         public virtual Sindicato Sindicato { get; set; }
